backend/requirements.txt
# marshmallow-sqlalchemy
# bleach
# Flask-Migrate
# flask-jwt-extended
flask-jwt-extended==4.6.0
#Email validation
email-validator==2.1.0.post1
# flask-limiter
#sumit req
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
pytest==7.4.2
pytest-flask==1.2.0

#### by Arpan
flask
flask-sqlalchemy
flask-socket
marshmallow
pytest
pytest-flask
pytest-mock